+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Moved all `import` statements to root `PowerModelsDistribution`
 - Fixed bug in `_calc_transformer_current_max_frto` where `sm_ub` should have been divided by `vmin`, NOT `vmax`
 - Updates `LPUBFDiagModel` description in docs